pyproject.toml
[tool.poetry]
name = "oz-django-team-2"
version = "0.1.0"
description = ""
authors = ["taewoo-dev <twin9774@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12"
django = "^5.1.1"
psycopg2 = "^2.9.9"
django-restframework = "^0.0.1"
djangorestframework-simplejwt = "^5.3.1"
requests = "^2.32.3"
drf-yasg = "^1.21.7"
python-dotenv = "^1.0.1"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
python_version = "3.12"
strict = true

[tool.mypy.django]
ignore_missing_imports = true

[tool.mypy.rest_framework]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "manage"
ignore_errors = true

[[tool.mypy.overrides]]
module = "rest_framework.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "rest_framework_simplejwt.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "rest_framework.generics.*"
ignore_missing_imports = true



[tool.django-stubs]
django_settings_module = "config.settings"

[tool.coverage.report]
# 테스트 80퍼 이상이면 통과
fail_under = 80

[tool.coverage.run]
# 테스트 제외
omit = [
  ".mypy_cache/**/*",
  "*/migrations/*",
  "*/*test*.py",
  "manage.py",
]

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
isort = "^5.13.2"
mypy = "^1.11.2"
django-stubs = "^5.0.4"
coverage = "^7.6.1"
django-extensions = "^3.2.3"
ipython = "^8.27.0"
djangorestframework-stubs = "^3.15.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
